requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
celery
redis
jinja2